Cleanup and simplify
